Overhaul entire randomwallpaper()
Far fewer processes are now required. Instead of using both `wget`
_and_ `curl`, now only `wget` is used. This can be changed, of course,
if one is preferred over the other.

There's now a `$file` and `$url` for cleaner code, easier reuse, and it
makes your life easier, should you desire or need to change either of
them.

Now `feh` will only set the background if `wget` succeeds in
downloading the file; this just makes this function a little more
robust, but won't account for all problems.

There was never a need for `head`. Always check the man pages.