Simplify `w.sh` for an easy one-liner
It occurred to me that this could be watered down to just the one line.
This would make for a nice alias, or something to have executed upon
logging into the user's account.

Ignore the previous commit's error message, in which I stated only
memory would be taken up, as, at least for me, a temporary file was
created in `/tmp`, which is still good, I think.